Improved UI logic for multiple menu screens
